+++ src/components/custom-field-input-single-choice/custom-field-input-single-choice.jsx
@@ -1,9 +1,11 @@
 import React, {
+  forwardRef,
+  useImperativeHandle,
   useRef,
   useState,
 } from 'react';
 import PropTypes from 'prop-types';
-import AbstractCustomField from '../__internal__/abstract-custom-field.jsx';
+import AbstractCustomField from '../__internal__/abstract-custom-field/abstract-custom-field.jsx';
 import Icon from '../icon/icon.jsx';
 import iconClear from '../../svgs/icon-clear-small.svg';
 import iconCaretDown from '../../svgs/icon-caret-down.svg';
@@ -11,10 +13,11 @@ import iconCaretDownDisabled from '../../svgs/icon-caret-down-disabled.svg';
 import styles from './custom-field-input-single-choice.css';
 import Listbox from '../listbox/listbox.jsx';
 import ListOption from '../list-option/list-option.jsx';
+import NoOptions from '../no-options/no-options.jsx';
 import useValidation from '../../hooks/use-validation.jsx';
 import useDropdownClose from '../../hooks/use-dropdown-close.js';
 
-export default function CustomFieldInputSingleChoice(props) {
+const CustomFieldInputSingleChoice = forwardRef(function CustomFieldInputSingleChoice(props, ref) {
   const [showOptions, setShowOptions] = useState(false);
   const [value, setValue] = useState(props.value);
   const [searchValue, setSearchValue] = useState(undefined);
@@ -90,8 +93,8 @@ export default function CustomFieldInputSingleChoice(props) {
     return props.choices.map(item => choices[item.id]);
   }
 
-  const listOptions = () => {
-    return getOptions()
+  const listOptions = (choices) => {
+    return choices
       .map(item => (
         <ListOption
           key={item.id}
@@ -128,6 +131,15 @@ export default function CustomFieldInputSingleChoice(props) {
     setShowOptions(true);
   }
 
+  useImperativeHandle(ref, () => ({
+    id: props.id,
+    get value() {
+      return value;
+    },
+  }));
+
+  const choices = getOptions();
+
   return (
     <div ref={wrapperRef} className={styles.container}>
       <AbstractCustomField
@@ -146,19 +158,21 @@ export default function CustomFieldInputSingleChoice(props) {
         value={searchValue || defaultValue}
       />
       { showOptions && (
-        <Listbox
-          className={styles.dropdown}
-          labelledBy={`${props.id}-label`}
-          onChange={onSelectionChange}
-          refs={refs}
-          value={value}
-        >
-          { listOptions() }
-        </Listbox>
-      ) }
+        choices.length === 0 ? (<NoOptions className={styles['no-options']} />) : (
+          <Listbox
+            className={styles.dropdown}
+            labelledBy={`${props.id}-label`}
+            onChange={onSelectionChange}
+            refs={refs}
+            value={value}
+          >
+            { listOptions(choices) }
+          </Listbox>
+        )
+      )}
     </div>
   );
-}
+});
 
 const ChoiceType = PropTypes.shape({
   id: PropTypes.string.isRequired,
@@ -184,3 +198,5 @@ CustomFieldInputSingleChoice.defaultProps = {
   value: undefined,
   errorText: undefined,
 };
+
+export default CustomFieldInputSingleChoice;